feat(settings): add General section for locale/household prefs
Location (weather) lived under Appearance, and Time zone + Week-start under
Calendars — the locale/household prefs were scattered in odd homes. Group all
three into a new 'General' settings section (near the top of the nav). Appearance
keeps theme/layout only; Calendars keeps calendar connections/hours. Pure move —
no logic changes.
